requirements_test.txt
black
coverage
pytest
flake8
mypy

paho-mqtt
redis
